backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
neo4j==5.15.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2

langchain==0.2.10
langchain-openai==0.1.9 
langchain-core==0.2.23
langgraph==0.1.24
python-dotenv==1.0.0